requirements.txt
fastapi
uvicorn[standard]
cryptography
qrcode[pil]
python-multipart
httpx
python-dotenv
